move cbor to own module
