+++ wallet-crypto/src/address.rs
@@ -5,139 +5,12 @@ extern crate rcw;
 use self::rcw::digest::Digest;
 use self::rcw::blake2b::Blake2b;
 use self::rcw::sha3::Sha3;
+use cbor;
 
 use hdwallet::{XPub};
 
-// internal mobule to encode the address metadata in cbor to
-// hash them.
-//
-mod cbor {
-    #[derive(Debug, PartialEq, PartialOrd, Eq, Ord, Copy, Clone)]
-    pub enum MajorType {
-        UINT,
-        NINT,
-        BYTES,
-        TEXT,
-        ARRAY,
-        MAP,
-        TAG,
-        T7
-    }
-
-    const INLINE_ENCODING : u8 = 24;
-
-    impl MajorType {
-        // serialize a major type in its highest bit form
-        fn to_byte(self) -> u8 {
-            use self::MajorType::*;
-            match self {
-                UINT  => 0b0000_0000,
-                NINT  => 0b0010_0000,
-                BYTES => 0b0100_0000,
-                TEXT  => 0b0110_0000,
-                ARRAY => 0b1000_0000,
-                MAP   => 0b1010_0000,
-                TAG   => 0b1100_0000,
-                T7    => 0b1110_0000
-            }
-        }
-    }
-
-    pub fn cbor_header(ty: MajorType, r: u8) -> u8 {
-        ty.to_byte() | r & 0x1f
-    }
-
-    pub fn cbor_uint_small(v: u8, buf: &mut Vec<u8>) {
-        assert!(v < INLINE_ENCODING);
-        buf.push(cbor_header(MajorType::UINT, v));
-    }
-
-    pub fn cbor_u8(v: u8, buf: &mut Vec<u8>) {
-        buf.push(cbor_header(MajorType::UINT, 24));
-        buf.push(v);
-    }
-
-    /// convenient macro to get the given bytes of the given value
-    ///
-    /// does all the job: Big Endian, bit shift and convertion
-    macro_rules! byte_slice {
-        ($value:ident, $shift:expr) => ({
-            ($value.to_be() >> $shift) as u8
-        });
-    }
-
-    pub fn write_u8(v: u8, buf: &mut Vec<u8>) {
-        write_header_u8(MajorType::UINT, v, buf);
-    }
-    pub fn write_u16(v: u16, buf: &mut Vec<u8>) {
-        write_header_u16(MajorType::UINT, v, buf);
-    }
-    pub fn write_u32(v: u32, buf: &mut Vec<u8>) {
-        write_header_u32(MajorType::UINT, v, buf);
-    }
-    pub fn write_u64(v: u64, buf: &mut Vec<u8>) {
-        write_header_u64(MajorType::UINT, v, buf);
-    }
-    pub fn write_header_u8(ty: MajorType, v: u8, buf: &mut Vec<u8>) {
-        buf.push(cbor_header(ty, 24));
-        buf.push(v);
-    }
-    pub fn write_header_u16(ty: MajorType, v: u16, buf: &mut Vec<u8>) {
-        buf.push(cbor_header(ty, 25));
-        buf.push(byte_slice!(v, 8));
-        buf.push(byte_slice!(v, 0));
-    }
-    pub fn write_header_u32(ty: MajorType, v: u32, buf: &mut Vec<u8>) {
-        buf.push(cbor_header(ty, 26));
-        buf.push(byte_slice!(v, 24));
-        buf.push(byte_slice!(v, 16));
-        buf.push(byte_slice!(v,  8));
-        buf.push(byte_slice!(v,  0));
-    }
-    pub fn write_header_u64(ty: MajorType, v: u64, buf: &mut Vec<u8>) {
-        buf.push(cbor_header(ty, 27));
-        buf.push(byte_slice!(v, 56));
-        buf.push(byte_slice!(v, 48));
-        buf.push(byte_slice!(v, 40));
-        buf.push(byte_slice!(v, 32));
-        buf.push(byte_slice!(v, 24));
-        buf.push(byte_slice!(v, 16));
-        buf.push(byte_slice!(v,  8));
-        buf.push(byte_slice!(v,  0));
-    }
-
-    pub fn write_length_encoding(ty: MajorType, nb_elems: usize, buf: &mut Vec<u8>) {
-        if nb_elems < (INLINE_ENCODING as usize) {
-            buf.push(cbor_header(ty, nb_elems as u8));
-        } else {
-            if nb_elems < 0x100 {
-                write_header_u8(ty, nb_elems as u8, buf);
-            } else if nb_elems < 0x10000 {
-                write_header_u16(ty, nb_elems as u16, buf);
-            } else if nb_elems < 0x100000000 {
-                write_header_u32(ty, nb_elems as u32, buf);
-            } else {
-                write_header_u64(ty, nb_elems as u64, buf);
-            }
-        }
-    }
-
-    pub fn cbor_bs(bs: &[u8], buf: &mut Vec<u8>) {
-        write_length_encoding(MajorType::BYTES, bs.len(), buf);
-        buf.extend_from_slice(bs)
-    }
-
-    pub fn cbor_array_start(nb_elems: usize, buf: &mut Vec<u8>) {
-        write_length_encoding(MajorType::ARRAY, nb_elems, buf);
-    }
-    pub fn cbor_map_start(nb_elems: usize, buf: &mut Vec<u8>) {
-        write_length_encoding(MajorType::MAP, nb_elems, buf);
-    }
-
-}
-
 mod hs_cbor {
-    use super::cbor::{cbor_array_start, write_length_encoding, MajorType};
+    use cbor::spec::{cbor_array_start, write_length_encoding, MajorType};
 
     pub fn sumtype_start(tag: u8, nb_values: usize, buf: &mut Vec<u8>) -> () {
         cbor_array_start(nb_values + 1, buf);
@@ -185,7 +58,7 @@ mod hs_cbor {
 
 mod hs_cbor_util {
     use hdwallet::{XPub};
-    use super::cbor::{cbor_bs};
+    use cbor::spec::{cbor_bs};
     pub fn cbor_xpub(pubk: &XPub, buf: &mut Vec<u8>) {
         cbor_bs(&pubk[..], buf);
     }
@@ -230,7 +103,7 @@ impl fmt::Display for DigestBlake2b {
 }
 impl ToCBOR for DigestBlake2b {
     fn encode(&self, buf: &mut Vec<u8>) {
-        cbor::cbor_bs(&self.0[..], buf)
+        cbor::spec::cbor_bs(&self.0[..], buf)
     }
 }
 
@@ -263,7 +136,7 @@ impl AddrType {
 }
 impl ToCBOR for AddrType {
     fn encode(&self, buf: &mut Vec<u8>) {
-        cbor::cbor_uint_small(self.to_byte(), buf);
+        cbor::spec::cbor_uint_small(self.to_byte(), buf);
     }
 }
 
@@ -312,7 +185,7 @@ impl ToCBOR for StakeDistribution {
                 si.encode(&mut vec);
             }
         };
-        cbor::cbor_bs(&vec, buf);
+        cbor::spec::cbor_bs(&vec, buf);
     }
 }
 
@@ -327,8 +200,8 @@ impl HDAddressPayload {
 impl ToCBOR for HDAddressPayload {
     fn encode(&self, buf: &mut Vec<u8>) {
         let mut vec = vec![];
-        cbor::cbor_bs(self.as_ref(), &mut vec);
-        cbor::cbor_bs(&vec         , buf);
+        cbor::spec::cbor_bs(self.as_ref(), &mut vec);
+        cbor::spec::cbor_bs(&vec         , buf);
     }
 }
 
@@ -354,11 +227,11 @@ impl Attributes {
 }
 impl ToCBOR for Attributes {
     fn encode(&self, buf: &mut Vec<u8>) {
-        cbor::cbor_map_start(2, buf);
+        cbor::spec::cbor_map_start(2, buf);
         // TODO
-        cbor::cbor_uint_small(0, buf);
+        cbor::spec::cbor_uint_small(0, buf);
         self.stake_distribution.encode(buf);
-        cbor::cbor_uint_small(1, buf);
+        cbor::spec::cbor_uint_small(1, buf);
         self.derivation_path.encode(buf);
     }
 }
